requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy>=2.0.32
openpyxl==3.1.2
Werkzeug==3.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0